Stab in the dark attempt to add new Bed Sediment Plotting Rule
https://internal.cida.usgs.gov/jira/browse/GCMON-268

- Plot sand D50 in cases where there is only one sample in a sample set,
with the mass of that one sample exceeding 20g.
- Please make sure the new rule is only applied when there is only one
sample in the sample set. Otherwise all plots of bed sediment will be
wrong.
- Colorado River at Eleven Mile Gage, 09522100 will have one more
display point at 3/23/2014 when this rule is in place

What I'm not sure about is that if this is going to occur every time, or
if there is other code in place which would limit this to only working
on sample sets of 1 sample. Otherwise, as the Criteria indicate, this
would muck up everything else. If the overall bed sediment average
doesn't function this way, kill my suggested commit and I'll throw it
back in the queue.